// lib/parser/types.ts
export interface MandateInfo {
  amount: string;
  nextDeductionDate: string;
  merchant: string;
  umn?: string;
  currency: string;
  pluginId: string;
  provider: string;
}